Split off `RoomCreationPreset` to its own file
So we can stop muddying up the actual config
